Move tests connected with executing block from LedgerSpec to BlockExecutionSpec
