+++ src/test/scala/io/iohk/ethereum/ledger/LedgerSpec.scala
@@ -14,7 +14,7 @@ import io.iohk.ethereum.consensus.validators.{ Validators, _ }
 import io.iohk.ethereum.crypto._
 import io.iohk.ethereum.domain._
 import io.iohk.ethereum.ledger.BlockExecutionError.{ ValidationAfterExecError, ValidationBeforeExecError }
-import io.iohk.ethereum.ledger.Ledger.{ BlockResult, PC, PR, VMImpl }
+import io.iohk.ethereum.ledger.Ledger.{ BlockResult, VMImpl }
 import io.iohk.ethereum.network.p2p.messages.PV62.BlockBody
 import io.iohk.ethereum.utils.MonetaryPolicyConfig
 import io.iohk.ethereum.vm._
@@ -32,49 +32,6 @@ import scala.concurrent.{ ExecutionContext, ExecutionContextExecutor }
 class LedgerSpec extends FlatSpec with PropertyChecks with Matchers with MockFactory with ScalaFutures {
 
   implicit val testContext: ExecutionContextExecutor = ExecutionContext.fromExecutor(Executors.newFixedThreadPool(4))
-
-  def createResult(
-    context: PC,
-    gasUsed: BigInt,
-    gasLimit: BigInt,
-    gasRefund: BigInt,
-    error: Option[ProgramError] = None,
-    returnData: ByteString = bEmpty,
-    logs: Seq[TxLogEntry] = Nil,
-    addressesToDelete: Set[Address] = Set.empty
-  ): PR = ProgramResult(
-    returnData = returnData,
-    gasRemaining = gasLimit - gasUsed,
-    world = context.world,
-    addressesToDelete = addressesToDelete,
-    logs = logs,
-    internalTxs = Nil,
-    gasRefund = gasRefund,
-    error = error
-  )
-
-  sealed trait Changes
-  case class UpdateBalance(amount: UInt256) extends Changes
-  case object IncreaseNonce extends Changes
-  case object DeleteAccount extends Changes
-
-  def applyChanges(stateRootHash: ByteString, blockchainStorages: BlockchainStorages, changes: Seq[(Address, Changes)]): ByteString = {
-    val initialWorld = BlockchainImpl(blockchainStorages).getWorldStateProxy(-1, UInt256.Zero, Some(stateRootHash),
-      noEmptyAccounts = false, ethCompatibleStorage = true)
-    val newWorld = changes.foldLeft[InMemoryWorldStateProxy](initialWorld){ case (recWorld, (address, change)) =>
-      change match {
-        case UpdateBalance(balanceIncrease) =>
-          val accountWithBalanceIncrease = recWorld.getAccount(address).getOrElse(Account.empty()).increaseBalance(balanceIncrease)
-          recWorld.saveAccount(address, accountWithBalanceIncrease)
-        case IncreaseNonce =>
-          val accountWithNonceIncrease = recWorld.getAccount(address).getOrElse(Account.empty()).increaseNonce()
-          recWorld.saveAccount(address, accountWithNonceIncrease)
-        case DeleteAccount =>
-          recWorld.deleteAccount(address)
-      }
-    }
-    InMemoryWorldStateProxy.persistState(newWorld).stateRootHash
-  }
 
   "Ledger" should "correctly calculate the total gas refund to be returned to the sender and paying for gas to the miner" in new TestSetup {
 
@@ -114,167 +71,6 @@ class LedgerSpec extends FlatSpec with PropertyChecks with Matchers with MockFac
       postTxWorld.getBalance(originAddress) shouldEqual (initialOriginBalance - balanceDelta)
       postTxWorld.getBalance(minerAddress) shouldEqual (initialMinerBalance + balanceDelta)
     }
-  }
-
-  it should "correctly change the nonce when executing a tx that results in contract creation" in new TestSetup {
-
-    val tx: Transaction =
-      defaultTx.copy(gasPrice = defaultGasPrice, gasLimit = defaultGasLimit, receivingAddress = None, payload = ByteString.empty)
-
-    val stx: SignedTransactionWithSender = SignedTransaction.sign(tx, originKeyPair, Some(blockchainConfig.chainId))
-
-    val header: BlockHeader = defaultBlockHeader.copy(beneficiary = minerAddress.bytes)
-
-    val postTxWorld: InMemoryWorldStateProxy =
-      consensus.blockPreparator.executeTransaction(stx.tx, stx.senderAddress, header, worldWithMinerAndOriginAccounts).worldState
-
-    postTxWorld.getGuaranteedAccount(originAddress).nonce shouldBe (initialOriginNonce + 1)
-  }
-
-  it should "correctly change the nonce when executing a tx that results in a message call" in new TestSetup {
-
-    val tx: Transaction = defaultTx.copy(
-      gasPrice = defaultGasPrice,
-      gasLimit = defaultGasLimit,
-      receivingAddress = Some(originAddress),
-      payload = ByteString.empty
-    )
-
-    val stx: SignedTransactionWithSender = SignedTransaction.sign(tx, originKeyPair, Some(blockchainConfig.chainId))
-
-    val header: BlockHeader = defaultBlockHeader.copy(beneficiary = minerAddress.bytes)
-
-    val postTxWorld: InMemoryWorldStateProxy =
-      consensus.blockPreparator.executeTransaction(stx.tx, stx.senderAddress, header, worldWithMinerAndOriginAccounts).worldState
-
-    postTxWorld.getGuaranteedAccount(originAddress).nonce shouldBe (initialOriginNonce + 1)
-  }
-
-  it should "correctly run executeBlockTransactions for a block without txs" in new BlockchainSetup {
-
-    val block = Block(validBlockHeader, validBlockBodyWithNoTxs)
-
-    override lazy val vm: VMImpl = new MockVM(c =>
-      createResult(context = c, gasUsed = defaultGasLimit, gasLimit = defaultGasLimit, gasRefund = 0)
-    )
-
-    val txsExecResult: Either[BlockExecutionError, BlockResult] = ledger.blockExecution.executeBlockTransactions(block)
-
-    assert(txsExecResult.isRight)
-    val BlockResult(resultingWorldState, resultingGasUsed, resultingReceipts) = txsExecResult.right.get
-    resultingGasUsed shouldBe 0
-    resultingReceipts shouldBe Nil
-    InMemoryWorldStateProxy.persistState(resultingWorldState).stateRootHash shouldBe validBlockHeader.stateRoot
-  }
-
-  it should "correctly run executeBlockTransactions for a block with one tx (that produces no errors)" in new BlockchainSetup {
-
-    val table: TableFor4[BigInt, Seq[TxLogEntry], Set[Address], Boolean] = Table[BigInt, Seq[TxLogEntry], Set[Address], Boolean](
-      ("gasLimit/gasUsed", "logs", "addressesToDelete", "txValidAccordingToValidators"),
-      (defaultGasLimit, Nil, Set.empty, true),
-      (defaultGasLimit / 2, Nil, defaultAddressesToDelete, true),
-      (2 * defaultGasLimit, defaultLogs, Set.empty, true),
-      (defaultGasLimit, defaultLogs, defaultAddressesToDelete, true),
-      (defaultGasLimit, defaultLogs, defaultAddressesToDelete, false)
-    )
-
-    forAll(table) { (gasLimit, logs, addressesToDelete, txValidAccordingToValidators) =>
-
-      val tx = validTx.copy(gasLimit = gasLimit)
-      val stx = SignedTransaction.sign(tx, originKeyPair, Some(blockchainConfig.chainId))
-
-      val blockHeader: BlockHeader = validBlockHeader.copy(gasLimit = gasLimit)
-      val blockBodyWithTxs: BlockBody = validBlockBodyWithNoTxs.copy(transactionList = Seq(stx.tx))
-      val block = Block(blockHeader, blockBodyWithTxs)
-
-      val mockValidators =
-        if (txValidAccordingToValidators) Mocks.MockValidatorsAlwaysSucceed
-        else Mocks.MockValidatorsAlwaysFail
-      val mockVm = new MockVM(c => createResult(
-        context = c,
-        gasUsed = UInt256(gasLimit),
-        gasLimit = UInt256(gasLimit),
-        gasRefund = UInt256.Zero,
-        logs = logs,
-        addressesToDelete = addressesToDelete
-      ))
-
-      // Beware we are not using `ledger`
-      val testLedger = newTestLedger(validators = mockValidators, vm = mockVm)
-
-      val txsExecResult = testLedger.blockExecution.executeBlockTransactions(block)
-
-      txsExecResult.isRight shouldBe txValidAccordingToValidators
-      if(txsExecResult.isRight){
-        val BlockResult(resultingWorldState, resultingGasUsed, resultingReceipts) = txsExecResult.right.get
-
-        val transaction = stx.tx.tx
-        // Check valid world
-        val minerPaymentForTxs = UInt256(transaction.gasLimit * transaction.gasPrice)
-        val changes = Seq(
-          originAddress -> IncreaseNonce,
-          originAddress -> UpdateBalance(-minerPaymentForTxs),          // Origin payment for tx execution and nonce increase
-          minerAddress -> UpdateBalance(minerPaymentForTxs)             // Miner reward for tx execution
-        ) ++ addressesToDelete.map(address => address -> DeleteAccount) // Delete all accounts to be deleted
-        val expectedStateRoot = applyChanges(validBlockParentHeader.stateRoot, blockchainStorages, changes)
-        expectedStateRoot shouldBe InMemoryWorldStateProxy.persistState(resultingWorldState).stateRootHash
-
-        // Check valid gasUsed
-        resultingGasUsed shouldBe stx.tx.tx.gasLimit
-
-        // Check valid receipts
-        resultingReceipts.size shouldBe 1
-        val Receipt(rootHashReceipt, gasUsedReceipt, logsBloomFilterReceipt, logsReceipt) = resultingReceipts.head
-        rootHashReceipt shouldBe HashOutcome(expectedStateRoot)
-        gasUsedReceipt shouldBe resultingGasUsed
-        logsBloomFilterReceipt shouldBe BloomFilter.create(logs)
-        logsReceipt shouldBe logs
-      }
-    }
-  }
-
-  it should "correctly run executeBlockTransactions for a block with one tx (that produces OutOfGas)" in new BlockchainSetup {
-
-    val blockBodyWithTxs: BlockBody = validBlockBodyWithNoTxs.copy(transactionList = Seq(validStxSignedByOrigin.tx))
-    val block = Block(validBlockHeader, blockBodyWithTxs)
-
-    override lazy val vm = new MockVM(c => createResult(
-      context = c,
-      gasUsed = UInt256(defaultGasLimit),
-      gasLimit = UInt256(defaultGasLimit),
-      gasRefund = UInt256.Zero,
-      logs = defaultLogs,
-      addressesToDelete = defaultAddressesToDelete,
-      error = Some(OutOfGas)
-    ))
-
-    val txsExecResult: Either[BlockExecutionError, BlockResult] = ledger.blockExecution.executeBlockTransactions(block)
-
-    assert(txsExecResult.isRight)
-    val BlockResult(resultingWorldState, resultingGasUsed, resultingReceipts) = txsExecResult.right.get
-
-    val transaction: Transaction = validStxSignedByOrigin.tx.tx
-    // Check valid world
-    val minerPaymentForTxs = UInt256(transaction.gasLimit * transaction.gasPrice)
-    val changes = Seq(
-      originAddress -> IncreaseNonce,
-      originAddress -> UpdateBalance(-minerPaymentForTxs),  // Origin payment for tx execution and nonce increase
-      minerAddress -> UpdateBalance(minerPaymentForTxs)     // Miner reward for tx execution
-    )
-    val expectedStateRoot: ByteString = applyChanges(validBlockParentHeader.stateRoot, blockchainStorages, changes)
-    expectedStateRoot shouldBe InMemoryWorldStateProxy.persistState(resultingWorldState).stateRootHash
-
-    // Check valid gasUsed
-    resultingGasUsed shouldBe transaction.gasLimit
-
-    // Check valid receipts
-    resultingReceipts.size shouldBe 1
-    val Receipt(rootHashReceipt, gasUsedReceipt, logsBloomFilterReceipt, logsReceipt) = resultingReceipts.head
-    rootHashReceipt shouldBe HashOutcome(expectedStateRoot)
-    gasUsedReceipt shouldBe resultingGasUsed
-    logsBloomFilterReceipt shouldBe BloomFilter.create(Nil)
-    logsReceipt shouldBe Nil
-
   }
 
   it should "correctly run executeBlock for a valid block without txs" in new BlockchainSetup {
